+++ database/migrations/2021_04_14_025214_create_job_invites_table.php
@@ -16,11 +16,11 @@ class CreateJobInvitesTable extends Migration
         Schema::create('job_invites', function (Blueprint $table) {
             $table->id();
             $table->text('message');
-            $table->unsignedInteger('job_id');
+            $table->unsignedBigInteger('job_id');
             $table->foreign('job_id')->references('id')->on('jobs')->onDelete('cascade');
-            $table->unsignedInteger('freelancer_id');
+            $table->unsignedBigInteger('freelancer_id');
             $table->foreign('freelancer_id')->references('id')->on('freelancers')->onDelete('cascade');
-            $table->unsignedInteger('client_id');
+            $table->unsignedBigInteger('client_id');
             $table->foreign('client_id')->references('id')->on('clients')->onDelete('cascade');
             $table->timestamps();
         });
